Introduce DistributionGroup model
This model is responsible for eventually associating a group of survey responses to a particular subset.  As an example, a user might create a `DistributionGroup` for just the local "Knights of Columbus" chapter.  The slug created on the `DistributionGroup` will be used to create a unique URL that will be sent to the "Knights of Columbus" chapter and when their `SurveyResponse` will eventually be associated to the `DistributionGroup`.